feat: ページ遷移の箇所をpathを使ったものに差し替え
