+++ components/pages/events/components/EventList/index.ts
@@ -1,0 +1,1 @@
+export { EventListPresenter as EventList } from "./EventListPresenter";
