--- components/pages/events/components/EventLists/index.ts
@@ -1,1 +1,0 @@
-export { EventListsPresenter as EventLists } from "./EventListsPresenter";
