+++ components/pages/events/components/UserList/index.ts
@@ -1,0 +1,1 @@
+export { UserListPresenter as UserList } from "./UserListPresenter";
