--- components/pages/events/components/UserLists/index.ts
@@ -1,1 +1,0 @@
-export { UserListsPresenter as UserLists } from "./UserListsPresenter";
